requirements.txt
beautifulsoup4
pandas
requests
selenium
python-dotenv

